Add background image and form section
This commit adds the background image for the form on the left hand side. It also adds the name for the company on an opaque div on top of the background image. This is on the lft hand side. On the right hand side, we have the form description, the form inputs, the submit button, and a log in button if the user has an account already.